+++ rastervision_pytorch_learner/rastervision/pytorch_learner/learner.py
@@ -797,7 +797,7 @@ class Learner(ABC):
         """
         if np.issubdtype(x.dtype, np.unsignedinteger):
             max_val = np.iinfo(x.dtype).max
-            x = x.astype(np.float32) / max_val
+            x = x.astype(float) / max_val
         return x
 
     def predict(self, x: Tensor, raw_out: bool = False) -> Any:
